Revert "healthd: BatteryMonitor: doLogValues: reduce log level to debug"
This reverts commit 90b0f82f771bf3fa40f916b144122faa1b6d37be [formerly aeccbb8f3a4f2a7256a6308b0216dcb88b4bff8c].

Reason for revert: needed for debug and investigations of battery and charging issues. The tooling depends on these lines (and have been for years). You should have these lines only on SOC changes and in the presence of a serious problem. Please do not remove logging without consulting with the BMS SW team first.

Change-Id: Ife0079a7e0c1499b22e843429832b019139c0d89

Former-commit-id: f08aeded81b56d0fdd23b1c102ed5a1871f3bc06